fix(seed): an activity links to the person who actually wrote it
A mail signed "Karoline Juettner" was filed against the Geschaeftsfuehrer.
Every email, call and meeting linked staff[0] -- the account's most senior
employee -- with no reference to who the body says wrote it.

That is not a near-miss. Asked who raised a complaint, an assistant reads the
signature in the body and answers with THAT name, while activity_link names
somebody else. The answer looks like a hallucination and is not: the data
contradicts itself, and the assistant is the thing that surfaces it. Three
accounts had it visibly -- Adsmasters, NETFORMIC, Reply -- and every other
account had the same shape, invisible only because nobody had asked yet.

demoActivity gains `person`. It names who at the CUSTOMER an activity was
with, and counterpartFor resolves it against that company's own staff.

A name the account does not employ is an ERROR, not a fallback to the senior
contact. A silent fallback here would restore the exact defect while looking
like it worked, which is how this shipped in the first place.

Without `person` the old behaviour stands, so the ~180 companies the dataset
never names keep the heuristic and their timelines.

records.go crossed the 500-line cap with the addition, so the file splits
along a seam that was already there: what HAPPENED on an account stays in
records.go, where an account STANDS -- lifecycle, rate card, offers -- moves
to standing.go. No logic moved with it.

Verified on an isolated stack seeded from the dataset: 45 activities carry a
person, every one of them employed at the company named, and every one named
in the body it is linked to. verify passes.
